Move portal stuff from levelmesh.h to portal.h
