+++ CMakeLists.txt
@@ -196,6 +196,7 @@ set( SOURCES
 	src/lightmap/glsl_vert.h
 	src/lightmap/glsl_frag_resolve.h
 	src/lightmap/renderdoc_app.h
+	src/lightmap/portal.h
 	src/math/mat.cpp
 	src/math/plane.cpp
 	src/math/angle.cpp
